Test Town Scene
Minor town scene created with roads added, Needs new ui elements and more buildings

+++ FinalTransmissionDemo/Assets/Scripts/TestSpecificScripts/SearchBuilding.cs
@@ -10,7 +10,7 @@ public class SearchBuilding : MonoBehaviour {
     private GameObject survivorOutside;
 
     [SerializeField]
-    private int BuildingSize =1 ;
+    private float BuildingSize =1 ;
 
 
     //Variables for Searching
